+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/home/InventoryScreen.kt
@@ -24,6 +24,7 @@ import androidx.navigation.NavHostController
 import com.app.arcabyolimpo.presentation.navigation.Screen
 import com.app.arcabyolimpo.presentation.screens.home.components.InventoryMainScreen
 import com.app.arcabyolimpo.presentation.screens.home.components.InventoryTopBar
+import com.app.arcabyolimpo.presentation.screens.productbatches.productBatchesList.ProductBatchesListScreen
 import com.app.arcabyolimpo.presentation.screens.supply.supplyList.SupplyListScreen
 import com.app.arcabyolimpo.presentation.ui.components.atoms.icons.NotificationIcon
 import com.app.arcabyolimpo.ui.theme.ArcaByOlimpoTheme
@@ -86,32 +87,13 @@ fun InventoryScreen(navController: NavHostController) {
                                 },
                             )
                         "product_batches" ->
-                            TopAppBar(
-                                title = {
-                                    Text(
-                                        "Lotes de Productos",
-                                        color = Color.White,
-                                        fontSize = 24.sp,
-                                        fontWeight = FontWeight.Bold,
-                                    )
+                            ProductBatchesListScreen(
+                                onBackClick = { selectedOption = null },
+                                onDetailClick = { id ->
+                                    navController.navigate(Screen.ProductBatchDetail.createRoute(id))
                                 },
-                                navigationIcon = {
-                                    IconButton(onClick = { selectedOption = null }) {
-                                        Icon(
-                                            imageVector = Icons.Default.ArrowBack,
-                                            contentDescription = "Regresar",
-                                            tint = Color.White,
-                                        )
-                                    }
-                                },
-                                colors =
-                                    TopAppBarDefaults.topAppBarColors(
-                                        containerColor = Color(0xFF040610),
-                                    ),
-                                actions = {
-                                    IconButton(onClick = { }) {
-                                        NotificationIcon()
-                                    }
+                                onAddClick = {
+                                    navController.navigate(Screen.ProductBatchRegister.route)
                                 },
                             )
                         "supplies" -> {