+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/productbatches/productBatchModify/ProductBatchModifyScreen.kt
@@ -81,9 +81,6 @@ fun ProductBatchModifyScreen(
                 colors = TopAppBarDefaults.topAppBarColors(containerColor = Background),
             )
         },
-        bottomBar = {
-            NavBar()
-        },
     ) { padding ->
         Column(
             modifier =